+++ lib.php
@@ -34,10 +34,6 @@ function theme_campus_process_css($css, $theme) {
         $customcss = null;
     }
 
-    if (\theme_campus\toolbox::get_setting('iconcoloursetting')) {
-        \theme_campus\toolbox::change_icons();
-    }
-
     $css = theme_campus_set_customcss($css, $customcss);
 
     return $css;