+++ settings.php
@@ -227,24 +227,6 @@ if ($ADMIN->fulltree) {
     $setting->set_updatedcallback('theme_reset_all_caches');
     $settingpage->add($setting);
 
-    // Icon colour setting 'setting' - to enable / disable the functionality.
-    $name = 'theme_campus/iconcoloursetting';
-    $title = get_string('iconcoloursetting', 'theme_campus');
-    $description = get_string('iconcoloursetting_desc', 'theme_campus');
-    $setting = new admin_setting_configcheckbox($name, $title, $description, 0);
-    $setting->set_updatedcallback('theme_reset_all_caches');
-    $settingpage->add($setting);
-
-    // Icon colour setting.
-    $name = 'theme_campus/iconcolour';
-    $title = get_string('iconcolour', 'theme_campus');
-    $description = get_string('iconcolourdesc', 'theme_campus');
-    $default = '#999999';
-    $previewconfig = null;
-    $setting = new admin_setting_configcolourpicker($name, $title, $description, $default, $previewconfig);
-    $setting->set_updatedcallback('theme_reset_all_caches');
-    $settingpage->add($setting);
-
     // Heading colour setting.
     $name = 'theme_campus/headingcolour';
     $title = get_string('headingcolour', 'theme_campus');